added inventory to user profile page

+++ Fairleih-iOS/Views/InventoryView/InventoryItemView.swift
@@ -33,14 +33,16 @@ struct InventoryItemView: View {
             .padding()
         }
         .contextMenu(ContextMenu(menuItems: {
-            Button(action: {
-                BackendClient.shared.deleteProduct(with: item?._id ?? "")
-            }, label: {
-                Text("Delete")
-                    .foregroundColor(.red)
-                Image(systemName: "trash")
-                    .foregroundColor(.red)
-            })
+            if(item?.user_id == UserObjectManager.shared.user?._id) {
+                Button(action: {
+                    BackendClient.shared.deleteProduct(with: item?._id ?? "")
+                }, label: {
+                    Text("Delete")
+                        .foregroundColor(.red)
+                    Image(systemName: "trash")
+                        .foregroundColor(.red)
+                })
+            }
         }))
     }
 }
